[skip travis]

+++ src/libsbn.cpp
@@ -11,15 +11,15 @@ namespace py = pybind11;
 class Matrix {
  public:
   Matrix(size_t rows, size_t cols) : m_rows(rows), m_cols(cols) {
-    m_data = new float[rows * cols];
+    m_data = new double[rows * cols];
   }
-  float *data() { return m_data; }
+  double *data() { return m_data; }
   size_t rows() const { return m_rows; }
   size_t cols() const { return m_cols; }
 
  private:
   size_t m_rows, m_cols;
-  float *m_data;
+  double *m_data;
 };
 
 Matrix make_matrix() {
@@ -27,7 +27,7 @@ Matrix make_matrix() {
   return x;
 }
 
-float get00(Matrix m) { return m.data()[0]; }
+double get00(Matrix m) { return m.data()[0]; }
 
 PYBIND11_MODULE(sbn, m) {
   m.doc() = "libsbn bindings";
@@ -45,14 +45,14 @@ PYBIND11_MODULE(sbn, m) {
   py::class_<Matrix>(m, "Matrix", py::buffer_protocol())
       .def_buffer([](Matrix &m) -> py::buffer_info {
         return py::buffer_info(
-            m.data(),                               /* Pointer to buffer */
-            sizeof(float),                          /* Size of one scalar */
-            py::format_descriptor<float>::format(), /* Python struct-style
+            m.data(),                                /* Pointer to buffer */
+            sizeof(double),                          /* Size of one scalar */
+            py::format_descriptor<double>::format(), /* Python struct-style
                                                        format descriptor */
-            2,                                      /* Number of dimensions */
-            {m.rows(), m.cols()},                   /* Buffer dimensions */
-            {sizeof(float) * m.rows(), /* Strides (in bytes) for each index */
-             sizeof(float)});
+            2,                                       /* Number of dimensions */
+            {m.rows(), m.cols()},                    /* Buffer dimensions */
+            {sizeof(double) * m.rows(), /* Strides (in bytes) for each index */
+             sizeof(double)});
       });
   m.def("make_matrix", &make_matrix, "test");
   m.def("get00", &get00, "test");
